Refactoring EmailSendUtil (funcionando 2).
