+++ src/main/java/com/polifono/util/EmailUtil.java
@@ -94,151 +94,6 @@ public class EmailUtil {
         DOMAIN_CORRECTIONS.put("@aiclod.com", "@icloud.com");
     }
 
-    public static void main(String[] args) {
-
-		/*String senha = "F123456";
-
-		if (validatePassword(senha)) {
-			System.out.println(senha + " é SENHA BOA");
-		}
-		else {
-			System.out.println(senha + " é SENHA RUIM");
-		}
-
-		senha = "123456";
-
-		if (validatePassword(senha)) {
-			System.out.println(senha + " é SENHA BOA");
-		}
-		else {
-			System.out.println(senha + " é SENHA RUIM");
-		}
-
-		senha = "AAAAAA";
-
-		if (validatePassword(senha)) {
-			System.out.println(senha + " é SENHA BOA");
-		}
-		else {
-			System.out.println(senha + " é SENHA RUIM");
-		}
-
-		senha = "F123 456";
-
-		if (validatePassword(senha)) {
-			System.out.println(senha + " é SENHA BOA");
-		}
-		else {
-			System.out.println(senha + " é SENHA RUIM");
-		}
-
-		senha = "F123*456";
-
-		if (validatePassword(senha)) {
-			System.out.println(senha + " é SENHA BOA");
-		}
-		else {
-			System.out.println(senha + " é SENHA RUIM");
-		}
-
-		senha = "F123(¨%*&*456";
-
-		if (validatePassword(senha)) {
-			System.out.println(senha + " é SENHA BOA");
-		}
-		else {
-			System.out.println(senha + " é SENHA RUIM");
-		}*/
-
-		/*String login = "flavio"; // OK
-
-		if (validateLogin(login)) {
-			System.out.println(login + " é LOGIN BOM");
-		}
-		else {
-			System.out.println(login + " é LOGIN RUIM");
-		}
-
-		login = "1234567890123456"; // OK
-
-		if (validateLogin(login)) {
-			System.out.println(login + " é LOGIN BOM");
-		}
-		else {
-			System.out.println(login + " é LOGIN RUIM");
-		}
-
-		login = "flavio10araujo"; // OK
-
-		if (validateLogin(login)) {
-			System.out.println(login + " é LOGIN BOM");
-		}
-		else {
-			System.out.println(login + " é LOGIN RUIM");
-		}
-
-		login = "flavio_araujo"; // ERROR
-
-		if (validateLogin(login)) {
-			System.out.println(login + " é LOGIN BOM");
-		}
-		else {
-			System.out.println(login + " é LOGIN RUIM");
-		}
-
-		login = "flavio araujo"; // ERROR
-
-		if (validateLogin(login)) {
-			System.out.println(login + " é LOGIN BOM");
-		}
-		else {
-			System.out.println(login + " é LOGIN RUIM");
-		}
-
-		login = "flavio*araujo"; // ERROR
-
-		if (validateLogin(login)) {
-			System.out.println(login + " é LOGIN BOM");
-		}
-		else {
-			System.out.println(login + " é LOGIN RUIM");
-		}
-
-		login = "flávio"; // ERROR
-
-		if (validateLogin(login)) {
-			System.out.println(login + " é LOGIN BOM");
-		}
-		else {
-			System.out.println(login + " é LOGIN RUIM");
-		}
-
-		login = "acão"; // ERROR
-
-		if (validateLogin(login)) {
-			System.out.println(login + " é LOGIN BOM");
-		}
-		else {
-			System.out.println(login + " é LOGIN RUIM");
-		}*/
-
-		/*System.out.println(avoidWrongDomain("test1@bol.com"));
-		System.out.println(avoidWrongDomain("test1@bol.com.br"));
-		System.out.println(avoidWrongDomain("test1@gmail.com"));
-		System.out.println(avoidWrongDomain("test1@gmail.com.br"));
-		System.out.println(avoidWrongDomain("test1@yahoo.com"));
-		System.out.println(avoidWrongDomain("test1@yahoo.com.br"));
-		System.out.println(avoidWrongDomain("test1@hotmail.com"));
-		System.out.println(avoidWrongDomain("test1@hotmail.com.br"));
-		System.out.println(avoidWrongDomain("test1@outlook.com"));
-		System.out.println(avoidWrongDomain("test1@outlook.com.br"));*/
-
-        System.out.println(avoidWrongDomain("test1@gmail.co"));
-        System.out.println(avoidWrongDomain("test1@gmail.comm"));
-        System.out.println(avoidWrongDomain("test1@gamail.com"));
-        System.out.println(avoidWrongDomain("test1@gamil.com"));
-    }
-
     /**
      * Validate an email.
      * Return true if the email is valid. False if the email is not valid.
